streamlit-app/requirements.txt
streamlit==1.32.0
pandas==2.2.0
Pillow==10.2.0
openai==1.12.0
requests==2.31.0
rembg==2.0.65
numpy==1.26.4